requirements.txt
streamlit
requests
python-dotenv
pandas
plotly
anthropic
sendgrid